// src/gen/models/v1/metric_type_dash_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file models/v1/metric_type_dash.proto (package dashboard.models.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file models/v1/metric_type_dash.proto.
 */
export const file_models_v1_metric_type_dash: GenFile = /*@__PURE__*/
  fileDesc("CiBtb2RlbHMvdjEvbWV0cmljX3R5cGVfZGFzaC5wcm90bxITZGFzaGJvYXJkLm1vZGVscy52MSKJAQoKTWV0cmljVHlwZRIWCg5tZXRyaWNfdHlwZV9pZBgBIAEoCRIRCglzZW5zb3JfaWQYAiABKAkSDAoEY29kZRgDIAEoCRIMCgRuYW1lGAQgASgJEg4KBnN5bWJvbBgFIAEoCRIRCgltaW5fdmFsdWUYBiABKAESEQoJbWF4X3ZhbHVlGAcgASgBQjpaOGdpdGh1Yi5jb20vTWVkaVN0YXRUZWNoL2Rhc2hib2FyZC1jbGllbnQvcGIvZ28vbW9kZWxzL3YxYgZwcm90bzM");

/**
 * @generated from message dashboard.models.v1.MetricType
 */
export type MetricType = Message<"dashboard.models.v1.MetricType"> & {
  /**
   * @generated from field: string metric_type_id = 1;
   */
  metricTypeId: string;

  /**
   * @generated from field: string sensor_id = 2;
   */
  sensorId: string;

  /**
   * @generated from field: string code = 3;
   */
  code: string;

  /**
   * @generated from field: string name = 4;
   */
  name: string;

  /**
   * @generated from field: string symbol = 5;
   */
  symbol: string;

  /**
   * @generated from field: double min_value = 6;
   */
  minValue: number;

  /**
   * @generated from field: double max_value = 7;
   */
  maxValue: number;
};

/**
 * Describes the message dashboard.models.v1.MetricType.
 * Use `create(MetricTypeSchema)` to create a new message.
 */
export const MetricTypeSchema: GenMessage<MetricType> = /*@__PURE__*/
  messageDesc(file_models_v1_metric_type_dash, 0);